+++ dev/install.r
@@ -1,9 +1,6 @@
 devel <- F
-pkgdown <- F
+pkgdown <- T
 
-# cpp_vendor() # run only when updating C++ headers
-devtools::clean_dll()
-cpp11::cpp_register()
 devtools::document()
 
 if (isTRUE(devel)) {